Restructure folders to have separate files for the model and controller of users, reviews and categories
