+++ categories/controller.js
@@ -1,0 +1,11 @@
+const { fetchCategories } = require("./model");
+
+const getCategories = (req, res, next) => {
+  fetchCategories()
+    .then((categories) => {
+      res.status(200).send({ categories });
+    })
+    .catch((err) => next(err));
+};
+
+module.exports = { getCategories };
