--- controller.js
@@ -1,26 +1,0 @@
-const { fetchCategories, fetchReviewById, fetchUsers } = require("./model");
-
-const getCategories = (req, res, next) => {
-  fetchCategories()
-    .then((categories) => {
-      res.status(200).send({ categories });
-    })
-    .catch((err) => next(err));
-};
-
-const getReviewById = (req, res, next) => {
-  const { review_id } = req.params;
-  fetchReviewById(review_id)
-    .then((review) => res.status(200).send({ review }))
-    .catch((err) => next(err));
-};
-
-const getUsers = (req, res, next) => {
-  fetchUsers()
-    .then((users) => {
-      res.status(200).send({ users });
-    })
-    .catch(next);
-};
-
-module.exports = { getCategories, getReviewById, getUsers };
