--- model.js
@@ -1,22 +1,0 @@
-const db = require("./db/connection");
-
-const fetchCategories = () => {
-  return db.query("SELECT * FROM categories").then(({ rows }) => rows);
-};
-
-const fetchReviewById = (review_id) => {
-  return db
-    .query("SELECT * FROM reviews WHERE review_id = $1", [review_id])
-    .then(({ rows }) => {
-      if (rows.length === 0) {
-        return Promise.reject({ status: 404, message: "ID not found" });
-      }
-      return rows[0];
-    });
-};
-
-const fetchUsers = () => {
-  return db.query("SELECT * FROM users").then(({ rows }) => rows);
-};
-
-module.exports = { fetchCategories, fetchReviewById, fetchUsers };
